set_tid_address should the tid of calling thread
set_tid_address is used by libpthread to get the tid for initial and
cloned thread (not gettid :|)
returning 0 will cause the thread's tid to be recorded as zero in TLS

Libpthread's rwlock utilize the tid for recording writer and reader
existance, 0 is used for indicating the lock is not locked.
Therefore, is a thread with tid=0 locks the rwlock, it will be come
ambiguous.

This can be reproduced with a setlocale(3) call, will trigger
rwlock_unlock to use elison dueto no __writer present.
This in term will trigger Intel RTM in an invalid region, causing GP(#0)